requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==0.28.0
python-dotenv==1.0.0
textblob==0.17.1